Create README.md
Project team 6 is an Android app we built as a team that utilized multi-threading. There wasn't a timer available at the time that enabled the user to have multiple synchronous timers going at the same time. This timer not only allowed the user to have multiple timers but they could also assign each timer a name, color and graphic.